+++ tamr_unify_client/project/attribute_mapping.py
@@ -1,4 +1,4 @@
-from tamr_unify_client.models.base_resource import BaseResource
+from tamr_unify_client.base_resource import BaseResource
 
 
 class AttributeMapping(BaseResource):
