+++ tamr_unify_client/project/attributemappingcollectionwithoutbase.py
@@ -2,7 +2,7 @@ import requests
 
 from tamr_unify_client import Client
 from tamr_unify_client.auth import UsernamePasswordAuth
-from tamr_unify_client.models.project.attributemappingwithoutbase import (
+from tamr_unify_client.project.attributemappingwithoutbase import (
     AttributeMappingNoBase,
 )
